Rename the invoker files for consistency with the examples that use them
